requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
cryptography==46.0.5
deep-translator==1.11.4
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
googletrans==4.0.0rc1
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.2
langdetect==1.0.9
MarkupSafe==3.0.2
motor==3.7.1
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
nltk==3.9.1
numpy==2.2.2
openai==2.24.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
patsy==1.0.1
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.2
pycparser==3.0
pydantic==2.11.7
pydantic-settings==2.13.1
pydantic_core==2.33.2
pyjokes==0.8.3
PyJWT==2.10.1
pymongo==4.16.0
pypdf==6.0.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyttsx3==2.98
pytz==2024.2
pywin32==310
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
rfc3986==1.5.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.3
statsmodels==0.14.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
twilio==9.7.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
